+++ __fxstatat.c
@@ -52,5 +52,6 @@ int __fxstatat(int ver, int fd, const char *path, struct stat *statbuf,
 	__debug("%s(fd: %i, path: '%s' -> '%s', ..., flags: 0x%x)\n", __func__,
 		fd, path, buf, flags);
 
+	__remove_at_empty_path_if_needed(buf, flags);
 	return __rootfxstatat(ver, fd, buf, statbuf, flags);
 }