+++ faccessat.c
@@ -48,5 +48,6 @@ int faccessat(int fd, const char *path, int mode, int flags)
 	__debug("%s(fd: %i, path: '%s' -> '%s', flags: 0x%x)\n", __func__, fd,
 		path, buf, flags);
 
+	__remove_at_empty_path_if_needed(buf, flags);
 	return next_faccessat(fd, buf, mode, flags);
 }