+++ unlinkat.c
@@ -50,5 +50,6 @@ int unlinkat(int fd, const char *path, int flags)
 	__debug("%s(fd: %d, path: '%s' -> '%s', flags: 0x%x)\n", __func__, fd,
 		path, buf, flags);
 
+	__remove_at_empty_path_if_needed(buf, flags);
 	return next_unlinkat(fd, buf, flags);
 }